# [cherry-pick] [branch-2.2] [Enhancement] check mem limit after fetch one chunk from storage engine (#5304)