--- a/be/src/exec/pipeline/olap_chunk_source.cpp
+++ b/be/src/exec/pipeline/olap_chunk_source.cpp
@@ -429,9 +429,9 @@ Status OlapChunkSource::_read_chunk_from_storage(RuntimeState* state, vectorized
     }
     SCOPED_TIMER(_scan_timer);
     do {
-        if (Status status = _prj_iter->get_next(chunk); !status.ok()) {
-            return status;
-        }
+        RETURN_IF_ERROR(state->check_mem_limit("read chunk from storage"));
+        RETURN_IF_ERROR(_prj_iter->get_next(chunk));
+
         TRY_CATCH_ALLOC_SCOPE_START()
 
         for (auto slot : _query_slots) {